+++ GuaranteedRate.Sextant/EncompassUtils/FieldUtils.cs
@@ -26,42 +26,91 @@ namespace GuaranteedRate.Sextant.EncompassUtils
      */
     public class FieldUtils
     {
-        private readonly IList<string> SIMPLE_FIELDS;
-        private readonly IList<string> MIDDLE_INDEXED;
-        private readonly IList<string> END_INDEXED;
-        private readonly IList<string> DOCUMENT_MULTI;
-        private readonly IList<string> MILESTONE_TASK_MULTI;
-        private readonly IList<string> NONE_MULTI;
-        private readonly IList<string> POST_CLOSING_CONDITION_MULTI;
-        private readonly IList<string> UNDERWRITING_MULTI;
-        private readonly IList<string> ROLE_MULTI_KEYS;
-        private readonly IList<string> MILESTONE_MULTI_KEYS;
+        private readonly ISet<string> SIMPLE_FIELDS;
+        private readonly ISet<string> MIDDLE_INDEXED;
+        private readonly ISet<string> END_INDEXED;
+        private readonly ISet<string> DOCUMENT_MULTI;
+        private readonly ISet<string> MILESTONE_TASK_MULTI;
+        private readonly ISet<string> NONE_MULTI;
+        private readonly ISet<string> POST_CLOSING_CONDITION_MULTI;
+        private readonly ISet<string> UNDERWRITING_MULTI;
+        private readonly ISet<string> ROLE_MULTI_KEYS;
+        private readonly ISet<string> MILESTONE_MULTI_KEYS;
 
-        private readonly IList<string> BORROWER_EMPLOYERS_MULTI_KEYS;
-        private readonly IList<string> CO_BORROWER_EMPLOYERS_MULTI_KEYS;
-        private readonly IList<string> BORROWER_RESIDENCES_MULTI_KEYS;
-        private readonly IList<string> CO_BORROWER_RESIDENCES_MULTI_KEYS;
-        private readonly IList<string> LIABILITIES_MULTI_KEYS;
-        private readonly IList<string> MORTGAGES_MULTI_KEYS;
+        private readonly ISet<string> BORROWER_EMPLOYERS_MULTI_KEYS;
+        private readonly ISet<string> CO_BORROWER_EMPLOYERS_MULTI_KEYS;
+        private readonly ISet<string> BORROWER_RESIDENCES_MULTI_KEYS;
+        private readonly ISet<string> CO_BORROWER_RESIDENCES_MULTI_KEYS;
+        private readonly ISet<string> LIABILITIES_MULTI_KEYS;
+        private readonly ISet<string> DEPOSITS_MULTI_KEYS;
+        private readonly ISet<string> MORTGAGES_MULTI_KEYS;
+        private readonly ISet<string> VESTING_MULTI_KEYS;
+
+        private readonly ISet<string> DISCLOSURES_MULTI_KEYS;
 
         private const string BORROWER_EMPLOYERS_STARTS = "BE";
         private const string CO_BORROWER_EMPLOYERS_STARTS = "CE";
         private const string BORROWER_RESIDENCES_STARTS = "BR";
         private const string CO_BORROWER_RESIDENCES_STARTS = "CR";
-        private const string FD_LIABILITIES_STARTS = "FD";
+        private const string FD_DEPOSITS_STARTS = "FD";
+        private const string DD_DEPOSITS_STARTS = "DD";
         private const string FL_LIABILITIES_STARTS = "FL";
+        private const string IRS_MORTGAGE_INFO_STARTS = "IR";
         private const string MORTGAGES_STARTS = "FM";
+        private const string VESTING_PARITIES_STARTS = "TR";
 
-        private readonly IDictionary<string, IList<string>> INDEX_MULTI_SORTER;
+        private const string DISCLOSURES_STARTS = "DISCLOSED";
+
+        /**
+         * Need to find indexes for:
+         *  AR00xx  - Tax info
+         *  AB00xx  - Affilitaed Businesses
+         *  HC00xx  - Home counsoling info
+         *  SP00xx  - Settlement provider
+         *  TA00xx  - Trust transaction info
+         */
+        private readonly IDictionary<string, ISet<string>> INDEX_MULTI_SORTER;
+        private readonly IDictionary<string, ISet<string>> INDEX_MULTI_SORTER_END;
+        private readonly ISet<string> UNKNOWN_KEYS;
 
         private static volatile FieldUtils encompassFields;
         private static object syncRoot = new Object();
 
         private static IList<FieldDescriptor> SELECTED_FIELDS;
 
-        public static readonly IList<string> BORROWER_PAIR_FIELDS = 
-            new List<string> { "4000", "4001", "4002", "4003", "4004", "4005", "4006", "4007", "65", "66", "97", 
-                                "1240", "FE0116", "FR0104", "FR0106", "FR0107", "FR0108", "1268" };
+        /**
+         * These *SEEM* to be the *Simple* fields that are affected by switching active borrower-pair.
+         * Have not found a good way to know which fields are affected by the active borrower-pair.
+         */
+        public static readonly ISet<string> BORROWER_PAIR_FIELDS =
+            new HashSet<string> { 
+                               "36", "37", "38", "39", "52", "53", "54", "60", "65", "66", "67", "68", "69", "70", "71", "97", "98", 
+                               "100", "101", "102", "103", "104", "105", "106", "107", "108", "109", "110", "111", "112", "113", 
+                               "114", "115", "116", "117", "118", "119", "120", "121", "122", "123", "124", "125", "126", 
+                               "144", "145", "146", "147", "148", "149", "150", "151", "152", "153", "154", "155", "156", "157", 
+                               "168", "169", "170", "171", "172", "173", "174", "175", "176", "177", "178", "179", "180", "181",
+                               "182", "183", "186", "188", "189", "191", "265", "266", "267", "268", "271", "272", "273",
+                               "403", "418", "461", "463", "464", "466", "467", "470", "471", "477", "478",
+                               "687", "900", "901", "902", "903", "904", "905", "906", "907", "908", "909", "910", "911", "912",
+                               "915", "919", "920", "921", "922", "923", "924", "933", "934", "936", "938", "940", "941", "942", 
+                               "943" , "981", "985", "1015", "1057", "1058", "1062", "1069", "1070", "1087", "1088", "1089", "1108",
+                               "1136", "1144", "1145", "1146", "1156", "1158", "1159", "168", "1169", "1170", "1171", 
+                               "1178", "1179", "1188", "1197", "1240", "1241", "1268", "1300", "1306", "1307", "1308", "1309", "1310",
+                               "1311", "1312", "1313", "1314", "1315", "1316", "1317", "1318", "1319", "1320", "1321", "1323", "1325",
+                               "1343", "1389", "1402", "1403", "1414", "1415", "1416", "1417", "1418", "1419", 
+                               "1450", "1452", "1480", "1484", "1490", "1502", "1519", "1520", "1521", "1522", "1523", "1524", "1525", 
+                               "1526", "1527", "1528", "1529", "1530", "1531", "1532", "1533", "1534", "1535", "1536", "1537", "1538", 
+                               "1758", "1759", "1815", "1816", "1817", "1818", "1819", "1820", "1868", "1873", "2849", "2850", 
+                               "4000", "4001", "4002", "4003", "4004", "4005", "4006", "4007", "4008", "4009", 
+                               "FE0102", "FE0103", "FE0104", "FE0105", "FE0106", "FE0107", "FE0109", "FE0110", "FE0113", "FE0115", 
+                               "FE0116", "FE0117", "FE0133", "FE0198", "FE0199", "FE0202", "FE0203", "FE0204", "FE0205", "FE0206", 
+                               "FE0207", "FE0209", "FE0210", "FE0213", "FE0215", "FE0216", "FE0217", "FE0233", "FE0298", "FE0299",
+                               "FR0104", "FR0106", "FR0107", "FR0108", "FR0112", "FR0115", "FR0124", "FR0198", "FR0199", "FR0204", 
+                               "FR0206", "FR0207", "FR0208", "FR0212", "FR0215", "FR0224", "FR0298", "FR0299", "FR0304", "FR0306", 
+                               "FR0307", "FR0308", "FR0312", "FR0315", "FR0324", "FR0398", "FR0399", "FR0404", "FR0406", "FR0407", 
+                               "FR0408", "FR0412", "FR0415", "FR0424", "FR0498", "FR0499" };
+
+
 
 
         public static void AddFieldCollection(FieldDescriptors fieldDescriptors) 
@@ -84,33 +133,45 @@ namespace GuaranteedRate.Sextant.EncompassUtils
 
         private FieldUtils()
         {
-            SIMPLE_FIELDS = new List<string>();
-            MIDDLE_INDEXED = new List<string>();
-            END_INDEXED = new List<string>();
-            DOCUMENT_MULTI = new List<string>();
-            MILESTONE_TASK_MULTI = new List<string>();
-            NONE_MULTI = new List<string>();
-            POST_CLOSING_CONDITION_MULTI = new List<string>();
-            UNDERWRITING_MULTI = new List<string>();
+            SIMPLE_FIELDS = new HashSet<string>();
+            MIDDLE_INDEXED = new HashSet<string>();
+            END_INDEXED = new HashSet<string>();
+            DOCUMENT_MULTI = new HashSet<string>();
+            MILESTONE_TASK_MULTI = new HashSet<string>();
+            NONE_MULTI = new HashSet<string>();
+            POST_CLOSING_CONDITION_MULTI = new HashSet<string>();
+            UNDERWRITING_MULTI = new HashSet<string>();
 
-            BORROWER_EMPLOYERS_MULTI_KEYS = new List<string>();
-            CO_BORROWER_EMPLOYERS_MULTI_KEYS = new List<string>();
-            BORROWER_RESIDENCES_MULTI_KEYS = new List<string>();
-            CO_BORROWER_RESIDENCES_MULTI_KEYS = new List<string>();
-            LIABILITIES_MULTI_KEYS = new List<string>();
-            MORTGAGES_MULTI_KEYS = new List<string>();
+            BORROWER_EMPLOYERS_MULTI_KEYS = new HashSet<string>();
+            CO_BORROWER_EMPLOYERS_MULTI_KEYS = new HashSet<string>();
+            BORROWER_RESIDENCES_MULTI_KEYS = new HashSet<string>();
+            CO_BORROWER_RESIDENCES_MULTI_KEYS = new HashSet<string>();
+            LIABILITIES_MULTI_KEYS = new HashSet<string>();
+            DEPOSITS_MULTI_KEYS = new HashSet<string>();
+            MORTGAGES_MULTI_KEYS = new HashSet<string>();
+            VESTING_MULTI_KEYS = new HashSet<string>();
 
-            INDEX_MULTI_SORTER = new Dictionary<string, IList<string>>();
+            INDEX_MULTI_SORTER = new Dictionary<string, ISet<string>>();
             INDEX_MULTI_SORTER.Add(BORROWER_EMPLOYERS_STARTS, BORROWER_EMPLOYERS_MULTI_KEYS);
             INDEX_MULTI_SORTER.Add(CO_BORROWER_EMPLOYERS_STARTS, CO_BORROWER_EMPLOYERS_MULTI_KEYS);
             INDEX_MULTI_SORTER.Add(BORROWER_RESIDENCES_STARTS, BORROWER_RESIDENCES_MULTI_KEYS);
             INDEX_MULTI_SORTER.Add(CO_BORROWER_RESIDENCES_STARTS, CO_BORROWER_RESIDENCES_MULTI_KEYS);
-            INDEX_MULTI_SORTER.Add(FD_LIABILITIES_STARTS, LIABILITIES_MULTI_KEYS);
+            INDEX_MULTI_SORTER.Add(FD_DEPOSITS_STARTS, DEPOSITS_MULTI_KEYS);
+            INDEX_MULTI_SORTER.Add(DD_DEPOSITS_STARTS, DEPOSITS_MULTI_KEYS);
             INDEX_MULTI_SORTER.Add(FL_LIABILITIES_STARTS, LIABILITIES_MULTI_KEYS);
             INDEX_MULTI_SORTER.Add(MORTGAGES_STARTS, MORTGAGES_MULTI_KEYS);
+            INDEX_MULTI_SORTER.Add(IRS_MORTGAGE_INFO_STARTS, MORTGAGES_MULTI_KEYS);
+            INDEX_MULTI_SORTER.Add(VESTING_PARITIES_STARTS, VESTING_MULTI_KEYS);
+
+            DISCLOSURES_MULTI_KEYS = new HashSet<string>();
+
+            INDEX_MULTI_SORTER_END = new Dictionary<string, ISet<string>>();
+            INDEX_MULTI_SORTER_END.Add(DISCLOSURES_STARTS, DISCLOSURES_MULTI_KEYS);
 
             ROLE_MULTI_KEYS = GetRoleMultiKeys();
             MILESTONE_MULTI_KEYS = GetMilestoneMultiKeys();
+
+            UNKNOWN_KEYS = new HashSet<string>();
 
             GetAllFieldIds();
         }
@@ -143,9 +204,9 @@ namespace GuaranteedRate.Sextant.EncompassUtils
             }
         }
 
-        private IList<string> GetRoleMultiKeys()
+        private ISet<string> GetRoleMultiKeys()
         {
-            IList<string> keys = new List<string>();
+            ISet<string> keys = new HashSet<string>();
             foreach (Role role in session.Loans.Roles)
             {
                 keys.Add(role.Name);
@@ -153,9 +214,9 @@ namespace GuaranteedRate.Sextant.EncompassUtils
             return keys;
         }
 
-        private IList<string> GetMilestoneMultiKeys()
+        private ISet<string> GetMilestoneMultiKeys()
         {
-            IList<string> keys = new List<string>();
+            ISet<string> keys = new HashSet<string>();
             foreach (Milestone m in session.Loans.Milestones)
             {
                 keys.Add(m.Name);
@@ -214,26 +275,47 @@ namespace GuaranteedRate.Sextant.EncompassUtils
                     switch (fieldDescriptor.InstanceSpecifierType)
                     {
                         case MultiInstanceSpecifierType.Index:
-                            //TODO: Still working the new way
-                            string key = fieldDescriptor.FieldID.Substring(0, 2);
-                            if (INDEX_MULTI_SORTER.Keys.Contains(key))
+                            string starts2 = fieldDescriptor.FieldID.Substring(0, 2);
+                            string starts3 = fieldDescriptor.FieldID.Substring(0, 3);
+                            /**
+                             * These 3 groups do not follow any known rules for extraction
+                             * have no descriptions, and seem to consist of dupe data.
+                             * 
+                             * Ignore them
+                             */
+                            if (starts2 != "LP" && starts3 != "FBE" && starts3 != "FCE")
                             {
-                                INDEX_MULTI_SORTER[key].Add(fieldDescriptor.FieldID);
-                            }
-                            
-                            //TODO: The old way
-                            if (fieldDescriptor.FieldID.Substring(2, 2) == "00")
-                            {
-                                MIDDLE_INDEXED.Add(fieldDescriptor.FieldID);
-                            }
-                            else if (fieldDescriptor.FieldID.Contains("00"))
-                            {
-                                //FIXME: Not sure how to handle these fields
-                                // MIDDLE_INDEXED.Add(fieldDescriptor.FieldID);
-                            }
-                            else
-                            {
-                                END_INDEXED.Add(fieldDescriptor.FieldID);
+                                //the < 30 is to skip DISCLOSEDGFE.Snapshot.NEWHUD.X100...
+                                if (fieldDescriptor.FieldID.Contains("00") && fieldDescriptor.FieldID.Length < 30)
+                                {
+                                    string key = fieldDescriptor.FieldID.Substring(0, 2);
+                                    if (INDEX_MULTI_SORTER.Keys.Contains(key))
+                                    {
+                                        INDEX_MULTI_SORTER[key].Add(fieldDescriptor.FieldID);
+                                    }
+                                    else
+                                    {
+                                        UNKNOWN_KEYS.Add(key);
+                                        MIDDLE_INDEXED.Add(fieldDescriptor.FieldID);
+                                    }
+                                }
+                                else
+                                {
+                                    bool matched = false;
+                                    foreach (string key in INDEX_MULTI_SORTER_END.Keys)
+                                    {
+                                        if (fieldDescriptor.FieldID.StartsWith(key))
+                                        {
+                                            INDEX_MULTI_SORTER_END[key].Add(fieldDescriptor.FieldID);
+                                            matched = true;
+                                            break;
+                                        }
+                                    }
+                                    if (!matched)
+                                    {
+                                        END_INDEXED.Add(fieldDescriptor.FieldID);
+                                    }
+                                }
                             }
                             break;
                         case MultiInstanceSpecifierType.Document:
@@ -259,9 +341,17 @@ namespace GuaranteedRate.Sextant.EncompassUtils
                     }
                 }
             }
+
+            /*
+            Console.WriteLine("UNKNOWN_KEYS\n------------");
+            foreach (string key in UNKNOWN_KEYS)
+            {
+                Console.WriteLine(key);
+            }
+            */
         }
 
-        private void UnrollMultiFieldIds(string fieldId, IList<string> keys)
+        private void UnrollMultiFieldIds(string fieldId, ISet<string> keys)
         {
             foreach (string key in keys)
             {
@@ -272,44 +362,89 @@ namespace GuaranteedRate.Sextant.EncompassUtils
         /**
          * These method are expected to be called repeatedly, so the results are cached.
          */
-        public static IList<string> SimpleFieldNames()
+        public static ISet<string> SimpleFieldNames()
         {
             return Instance.SIMPLE_FIELDS;
         }
 
-        public static IList<string> MiddleIndexMulti()
+        public static ISet<string> MiddleIndexMulti()
         {
             return Instance.MIDDLE_INDEXED;
         }
 
-        public static IList<string> EndIndexMulti()
+        public static ISet<string> EndIndexMulti()
         {
             return Instance.END_INDEXED;
         }
 
-        public static IList<string> DocumentMulti()
+        public static ISet<string> DocumentMulti()
         {
             return Instance.DOCUMENT_MULTI;
         }
 
-        public static IList<string> MilestoneTaskMulti()
+        public static ISet<string> MilestoneTaskMulti()
         {
             return Instance.MILESTONE_TASK_MULTI;
         }
 
-        public static IList<string> PostClosingMulti()
+        public static ISet<string> PostClosingMulti()
         {
             return Instance.POST_CLOSING_CONDITION_MULTI;
         }
 
-        public static IList<string> RoleMultiKeys()
+        public static ISet<string> RoleMultiKeys()
         {
             return Instance.ROLE_MULTI_KEYS;
         }
 
-        public static IList<string> UnderwritingMulti()
+        public static ISet<string> UnderwritingMulti()
         {
             return Instance.UNDERWRITING_MULTI;
+        }
+
+        public static ISet<string> BorrowerEmployers()
+        {
+            return Instance.BORROWER_EMPLOYERS_MULTI_KEYS;
+        }
+
+        public static ISet<string> CoBorrowerEmployers()
+        {
+            return Instance.CO_BORROWER_EMPLOYERS_MULTI_KEYS;
+        }
+
+        public static ISet<string> BorrowerResidences()
+        {
+            return Instance.BORROWER_RESIDENCES_MULTI_KEYS;
+        }
+
+        public static ISet<string> CoBorrowerResidences()
+        {
+            return Instance.CO_BORROWER_RESIDENCES_MULTI_KEYS;
+        }
+
+        public static ISet<string> LiabilitiesMulti()
+        {
+            return Instance.LIABILITIES_MULTI_KEYS;
+        }
+
+        public static ISet<string> DepostisMulti()
+        {
+            return Instance.DEPOSITS_MULTI_KEYS;
+        }
+
+        public static ISet<string> MortgagesMulti()
+        {
+            return Instance.MORTGAGES_MULTI_KEYS;
+        }
+
+        public static ISet<string> VestingPartiesMulti()
+        {
+            return Instance.VESTING_MULTI_KEYS;
+        }
+
+        public static ISet<string> DisclosureMulti()
+        {
+            return Instance.DISCLOSURES_MULTI_KEYS;
         }
 
         /**