+++ GuaranteedRate.Util.IndexFields/Properties/AssemblyInfo.cs
@@ -5,12 +5,12 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("GuaranteedRate.Sextant")]
-[assembly: AssemblyDescription("Tools to navigate Encompass")]
+[assembly: AssemblyTitle("GuaranteedRate.Util.IndexFields")]
+[assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Guaranteed Rate")]
-[assembly: AssemblyProduct("GuaranteedRate.Sextant")]
-[assembly: AssemblyCopyright("Copyright © Guaranteed Rate 2015-2016")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("GuaranteedRate.Util.IndexFields")]
+[assembly: AssemblyCopyright("Copyright ©  2016")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("f279cfe6-0aa0-412f-8d7a-c933911ecdc3")]
+[assembly: Guid("91cb43df-5a4d-49eb-8d12-147a34a5bad3")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.9.1.0")]
-[assembly: AssemblyFileVersion("1.9.1.0")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]